Update CSS link paths
Previously the project failed to find the css files to md5 hash them
because they were absolute paths. If you simply change the paths to
relative paths, then non-root level urls fail to load. So this change
makes the md5 hash take the relative path, but outputs an absolute one
to the browser.