Remove menu border hidden style on hover
Perhaps this is just unfinished, but right now when you hover over menu
items, the right border disappears which changes the width of the
element and causes all menu items after the hovered one to reflow. It
looks fine to me without hiding it. (My gut is that this was borrowed
from another menu where this made sense.)